Move types outside of `rpc-request` file
