+++ packages/rpc-spec/src/index.ts
@@ -1,5 +1,4 @@
 export * from './rpc';
 export * from './rpc-api';
-export * from './rpc-request';
 export * from './rpc-shared';
 export * from './rpc-transport';
