+++ packages/rpc-spec/src/rpc-api.ts
@@ -1,11 +1,15 @@
 import { Callable } from '@solana/rpc-spec-types';
 
-import { RpcApiRequestPlan } from './rpc-request';
-
 export type RpcApiConfig = Readonly<{
     parametersTransformer?: <T extends unknown[]>(params: T, methodName: string) => unknown;
     responseTransformer?: <T>(response: unknown, methodName?: string) => T;
 }>;
+
+export type RpcApiRequestPlan<TResponse> = {
+    methodName: string;
+    params: unknown;
+    responseTransformer?: (response: unknown, methodName: string) => TResponse;
+};
 
 export type RpcApi<TRpcMethods> = {
     [MethodName in keyof TRpcMethods]: RpcReturnTypeMapper<TRpcMethods[MethodName]>;
