--- packages/rpc-spec/src/rpc-request.ts
@@ -1,13 +1,0 @@
-export type RpcApiRequestPlan<TResponse> = {
-    methodName: string;
-    params: unknown;
-    responseTransformer?: (response: unknown, methodName: string) => TResponse;
-};
-
-export type PendingRpcRequest<TResponse> = {
-    send(options?: RpcSendOptions): Promise<TResponse>;
-};
-
-export type RpcSendOptions = Readonly<{
-    abortSignal?: AbortSignal;
-}>;
